refactor: 마이페이지 제거 및 SignIn, SignUp 페이지 추가
